+++ src/main/kotlin/me/fru1t/lwnetmon/lwNetMonApplication.kt
@@ -17,12 +17,12 @@ fun main(args: Array<String>) {
   // 1 interface |2bytes 3packets 4errs 5drop 6fifo 7frame 8compressed 9multicast|10bytes 11packets 12errs 13drop 14fifo 15colls 16carrier 17compressed
   val format = Pattern.compile("\\s*([^:]+): (\\d+)\\s+(\\d+)\\s+(\\d+)\\s+(\\d+)\\s+(\\d+)\\s+(\\d+)\\s+(\\d+)\\s+(\\d+)\\s+(\\d+)\\s+(\\d+)\\s+(\\d+)\\s+(\\d+)\\s+(\\d+)\\s+(\\d+)\\s+(\\d+)\\s+(\\d+)")
   val lInterface = "enp30s0"
-  var lastRx = 0
-  var lastTx = 0
-  var currentRx = 0
-  var currentTx = 0
-  var deltaRx = 0
-  var deltaTx = 0
+  var lastRx = 0L
+  var lastTx = 0L
+  var currentRx = 0L
+  var currentTx = 0L
+  var deltaRx = 0L
+  var deltaTx = 0L
 
   val baselineTimestampMs = System.nanoTime() / 1000 / 1000
   var currentLoopTime: Long
@@ -41,8 +41,8 @@ fun main(args: Array<String>) {
         return@forEachLine
       }
 
-      currentRx = matcher.group(2).toInt()
-      currentTx = matcher.group(10).toInt()
+      currentRx = matcher.group(2).toLong()
+      currentTx = matcher.group(10).toLong()
       deltaRx = (currentRx - lastRx)
       deltaTx = (currentTx - lastTx)
 
@@ -64,7 +64,7 @@ fun main(args: Array<String>) {
   }
 }
 
-fun toHumanReadableString(bytes: Int): String {
+fun toHumanReadableString(bytes: Long): String {
   if (bytes < 1024) {
     return "$bytes B/s"
   }
